pyproject.toml
[tool.poetry]
name = "automacao-python"
version = "0.1.0"
description = ""
authors = ["Jhwnyor <franciscohp0007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.25.2"
pydantic = "^1.9.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
